+++ lib/api/presenters/root_presenter.rb
@@ -63,6 +63,13 @@ module Api
         }
       end
 
+      link :data do |opts|
+        {
+          href: "#{base_url(opts)}/api/data/{channel_id,team_id}",
+          templated: true
+        }
+      end
+
       %i[user channel team round sup].each do |model|
         link model do |opts|
           {