+++ lib/commands/help.rb
@@ -23,6 +23,7 @@ module SlackSup
         set api token              - require an access token in the X-Access-Token header for API access
         unset api token            - don't require an access token for API access
         rotate api token           - rotate the token required for API access
+        data                       - get a .zip of the team's data
 
         In a Channel
         ------------
@@ -30,6 +31,7 @@ module SlackSup
         rounds [n]                 - channel stats for the last n rounds, default is 3
         next                       - time of next round
         set                        - show all channel settings
+        data                       - get DMed a link to a .zip of the channel data
 
         In a Channel (Admins)
         --------------