+++ lib/models/channel_stats.rb
@@ -1,6 +1,7 @@
 class ChannelStats
   include ActiveModel::Model
   include SlackSup::Models::Mixins::Pluralize
+  include SlackSup::Models::Mixins::Export
 
   attr_accessor :rounds_count, :sups_count, :users_in_sups_count, :users_opted_in_count, :users_count, :outcomes, :channel