+++ lib/models/mixins/short_lived_token.rb
@@ -6,7 +6,7 @@ module SlackSup
 
         def short_lived_token
           JWT.encode({ dt: Time.now.utc.to_i, nonce: SecureRandom.hex }, token)
-          end
+        end
 
         def short_lived_token_valid?(short_lived_token, dt = 30.minutes)
           return false unless short_lived_token