+++ lib/models/round_stats.rb
@@ -1,6 +1,7 @@
 class RoundStats
   include ActiveModel::Model
   include SlackSup::Models::Mixins::Pluralize
+  include SlackSup::Models::Mixins::Export
 
   attr_accessor :round, :sups_count, :users_in_sups_count, :outcomes