+++ spec/models/channel_spec.rb
@@ -574,4 +574,76 @@ describe Channel do
       ].join("\n")
     end
   end
+
+  describe '#export' do
+    let(:channel) { Fabricate(:channel) }
+
+    include_context 'uses temp dir'
+
+    before do
+      allow(channel).to receive(:sync!)
+      allow(channel).to receive(:inform!)
+      channel.sup!
+      channel.export!(tmp)
+    end
+
+    %w[channel rounds sups stats users].each do |csv|
+      it "creates #{csv}.csv" do
+        expect(File.exist?(File.join(tmp, "#{csv}.csv"))).to be true
+      end
+    end
+
+    it 'creates rounds subfolders' do
+      expect(Dir.exist?(File.join(tmp, 'rounds'))).to be true
+      expect(File.exist?(File.join(tmp, 'rounds', channel.rounds.first.ran_at.strftime('%F'), 'round.csv'))).to be true
+    end
+
+    context 'channel.csv' do
+      let(:csv) { CSV.read(File.join(tmp, 'channel.csv'), headers: true) }
+
+      it 'generates csv' do
+        expect(csv.headers).to eq(
+          %w[
+            id
+            enabled
+            channel_id
+            created_at
+            updated_at
+            sup_wday
+            sup_followup_wday
+            sup_day
+            sup_tz
+            sup_time_of_day
+            sup_time_of_day_s
+            sup_every_n_weeks
+            sup_size
+          ]
+        )
+        row = csv[0]
+        expect(row['channel_id']).to eq channel.channel_id
+      end
+    end
+  end
+
+  describe '#export_zip!' do
+    let(:channel) { Fabricate(:channel) }
+
+    include_context 'uses temp dir'
+
+    before do
+      allow(channel).to receive(:sync!)
+      allow(channel).to receive(:inform!)
+      channel.sup!
+      channel.export!(tmp)
+    end
+
+    context 'zip' do
+      let!(:zip) { channel.export_zip!(tmp) }
+
+      it 'exists' do
+        expect(File.exist?(zip)).to be true
+        expect(File.size(zip)).not_to eq 0
+      end
+    end
+  end
 end