+++ spec/models/sup_spec.rb
@@ -141,6 +141,32 @@ describe Sup do
             expect(sup.send(:intro_message)).to be_nil
           end
         end
+
+        describe '#export' do
+          include_context 'uses temp dir'
+
+          before do
+            sup.export!(tmp)
+          end
+
+          %w[sup].each do |csv|
+            it "creates #{csv}.csv" do
+              expect(File.exist?(File.join(tmp, "#{csv}.csv"))).to be true
+            end
+          end
+
+          context 'sup.csv' do
+            let(:csv) { CSV.read(File.join(tmp, 'sup.csv'), headers: true) }
+
+            it 'generates csv' do
+              expect(csv.headers).to eq(%w[id outcome created_at updated_at captain_user_name users])
+              row = csv[0]
+              expect(row['outcome']).to be_nil
+              expect(row['users']).to eq sup.users.map(&:user_name).join("\n")
+              expect(row['captain_user_name']).to eq sup.captain.user_name
+            end
+          end
+        end
       end
     end
   end